Import of UFPS1
Import of UFPS1 and fixing some warnings.
UFPS will be the new base of the PlayerController & Interaction system.
There may be some import warnings whan first opening the project. These can be safely ignored, and they will popup just once.

PLEASE NOTE:
You are not allowed to reuse this AddOn in another project, since it is licensed to Doggielino. You have to obtain your own license if you want to do that.